+++ packages/mana/base.go
@@ -37,11 +37,6 @@ func (bm *BaseMana) updateBM2(n time.Duration) {
 	bm.BaseMana2 = bm.BaseMana2 * math.Pow(math.E, -Decay*n.Seconds())
 }
 
-// GetBM2 computes and returns the EBM2 based on the value and duration specified.
-func GetBM2(value float64, n time.Duration) float64 {
-	return value * math.Pow(math.E, -Decay*n.Seconds())
-}
-
 func (bm *BaseMana) updateEBM2(n time.Duration) {
 	if emaCoeff2 != Decay {
 		bm.EffectiveBaseMana2 = math.Pow(math.E, -emaCoeff2*n.Seconds())*bm.EffectiveBaseMana2 +
